requirements.txt
tensorflow>=2.15.0
streamlit==1.35.0
numpy==1.26.4
pillow==10.3.0